+++ packages/react/src/Dialog/index.tsx
@@ -1,0 +1,3 @@
+export {Dialog} from './Dialog'
+
+export type {DialogButtonProps, DialogHeaderProps, DialogHeight, DialogProps, DialogWidth} from './Dialog'